requirements.txt
asgiref==3.8.1
bleach==6.1.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
crispy-bootstrap5==0.7
cryptography==42.0.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.15
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.2
django-db-geventpool==4.0.6
django-summernote==0.8.20.0
djangorestframework==3.15.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
Pillow==8.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==1.26.19
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==5.3.0
zipp==3.19.0
